Remove some translations Rename Privacy settings
