+++ app/src/main/java/com/hippo/ehviewer/Settings.java
@@ -578,7 +578,7 @@ public class Settings {
     /********************
      ****** Privacy and Security
      ********************/
-    public static final String KEY_SEC_SECURITY = "enabled_secure";
+    public static final String KEY_SEC_SECURITY = "enable_secure";
     public static final boolean VALUE_SEC_SECURITY = false;
 
     public static boolean getEnabledSecurity() {
