--- app/src/main/java/com/hippo/ehviewer/ui/fragment/SecFragment.java
@@ -1,30 +1,0 @@
-package com.hippo.ehviewer.ui.fragment;
-
-import android.os.Bundle;
-import android.preference.Preference;
-import android.preference.PreferenceFragment;
-import android.text.TextUtils;
-
-import com.hippo.ehviewer.R;
-import com.hippo.ehviewer.Settings;
-
-/**
- * Created by Mo10 on 2018/2/10.
- */
-
-public class SecFragment extends PreferenceFragment {
-    private static final String KEY_PATTERN_PROTECTION = "pattern_protection";
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        addPreferencesFromResource(R.xml.secure_settings);
-    }
-    @Override
-    public void onResume() {
-        super.onResume();
-        Preference patternProtection = findPreference(KEY_PATTERN_PROTECTION);
-        patternProtection.setSummary(TextUtils.isEmpty(Settings.getSecurity()) ?
-                R.string.settings_advanced_pattern_protection_not_set :
-                R.string.settings_advanced_pattern_protection_set);
-    }
-}
