+++ app/src/main/res/values-zh-rHK/strings.xml
@@ -480,9 +480,9 @@
     <string name="settings_advanced_clear_memory_cache">清理內存緩存</string>
     <string name="settings_advanced_clear_memory_cache_summary">清理縮略圖、畫廊信息的內存緩存</string>
     <string name="settings_advanced_read_cache_size">閲讀緩存大小</string>
-    <string name="settings_advanced_pattern_protection_title">圖案保護</string>
-    <string name="settings_advanced_pattern_protection_not_set">未設置圖案保護</string>
-    <string name="settings_advanced_pattern_protection_set">已設置圖案保護</string>
+    <string name="settings_privacy_pattern_protection_title">圖案保護</string>
+    <string name="settings_privacy_pattern_protection_not_set">未設置圖案保護</string>
+    <string name="settings_privacy_pattern_protection_set">已設置圖案保護</string>
     <string name="settings_advanced_export_data">導出數據</string>
     <string name="settings_advanced_export_data_summary">保存數據至外置存儲器，例如下載列表，快速搜索列表</string>
     <string name="settings_advanced_export_data_to">已導出數據至 %s</string>
@@ -504,8 +504,8 @@
     <string name="settings_about_donate_summary">支付寶：seven332$163.com</string>
     <string name="settings_about_donate_message">你的捐贈能幫助我開發更多更好的應用。\n\n支付寶：seven332$163.com</string>
     <string name="settings_about_donate_toast">郵箱地址已複製到剪切板</string>
-    <string name="settings_about_enable_analytics">啟用數據統計</string>
-    <string name="settings_about_about_analytics">關於數據統計</string>
+    <string name="settings_privacy_enable_analytics">啟用數據統計</string>
+    <string name="settings_privacy_about_analytics">關於數據統計</string>
     <string name="settings_about_version">版本號</string>
     <string name="settings_about_check_for_updates">檢查更新</string>
     <string name="settings_about_beta_update_channel">更新測試版</string>
@@ -627,8 +627,8 @@
     <string name="download_archive_failure">無法下載壓縮包</string>
     <string name="archive">壓縮包</string>
     <string name="download_archive_failure_no_hath">下載壓縮包需要 H@H 客户端</string>
-    <string name="settings_sec">隱私和安全</string>
-    <string name="settings_sec_secure">不允許荧幕抓取</string>
-    <string name="settings_sec_secure_summary">啟用後，將不能截取該應用的螢幕截圖，同時，將不會在系統任務切換器中顯示該應用的內容預覽\n\n重新啟動應用以生效此更改</string>
+    <string name="settings_privacy">隱私和安</string>
+    <string name="settings_privacy_secure">不允許荧幕抓取</string>
+    <string name="settings_privacy_secure_summary">啟用後，將不能截取該應用的螢幕截圖，同時，將不會在系統任務切換器中顯示該應用的內容預覽\n\n重新啟動應用以生效此更改</string>
 
 </resources>
