+++ app/src/main/res/values-zh-rTW/strings.xml
@@ -480,9 +480,9 @@
     <string name="settings_advanced_clear_memory_cache">清除記憶體快取</string>
     <string name="settings_advanced_clear_memory_cache_summary">清除縮圖、圖庫詳細資料的記憶體快取</string>
     <string name="settings_advanced_read_cache_size">閱讀快取大小</string>
-    <string name="settings_advanced_pattern_protection_title">繪圖解鎖</string>
-    <string name="settings_advanced_pattern_protection_not_set">尚未設定繪圖解鎖</string>
-    <string name="settings_advanced_pattern_protection_set">已設定繪圖解鎖</string>
+    <string name="settings_privacy_pattern_protection_title">繪圖解鎖</string>
+    <string name="settings_privacy_pattern_protection_not_set">尚未設定繪圖解鎖</string>
+    <string name="settings_privacy_pattern_protection_set">已設定繪圖解鎖</string>
     <string name="settings_advanced_export_data">匯出檔案</string>
     <string name="settings_advanced_export_data_summary">儲存資料至外接儲存設備，例如下載列表，快速搜尋列表</string>
     <string name="settings_advanced_export_data_to">已匯出資料至 %s</string>
@@ -504,8 +504,8 @@
     <string name="settings_about_donate_summary">支付寶：seven332$163.com</string>
     <string name="settings_about_donate_message">你的贊助能幫助我開發更多更好的軟體。\n\n支付寶：seven332$163.com</string>
     <string name="settings_about_donate_toast">已將電子信箱地址複製到剪貼簿</string>
-    <string name="settings_about_enable_analytics">啟用資料統計</string>
-    <string name="settings_about_about_analytics">關於資料統計</string>
+    <string name="settings_privacy_enable_analytics">啟用資料統計</string>
+    <string name="settings_privacy_about_analytics">關於資料統計</string>
     <string name="settings_about_version">版本號碼</string>
     <string name="settings_about_check_for_updates">檢查更新</string>
     <string name="settings_about_beta_update_channel">更新測試版</string>
@@ -627,8 +627,8 @@
     <string name="download_archive_failure">下載失敗</string>
     <string name="archive">壓縮檔</string>
     <string name="download_archive_failure_no_hath">下載壓縮檔需要 H@H 客戶端</string>
-    <string name="settings_sec">隱私和安全</string>
-    <string name="settings_sec_secure">不允許荧幕抓取</string>
-    <string name="settings_sec_secure_summary">啟用後，將不能截取該應用的螢幕截圖，同時，將不會在系統任務切換器中顯示該應用的內容預覽\n\n重新啟動應用以生效此更改</string>
+    <string name="settings_privacy">隱私</string>
+    <string name="settings_privacy_secure">不允許荧幕抓取</string>
+    <string name="settings_privacy_secure_summary">啟用後，將不能截取該應用的螢幕截圖，同時，將不會在系統任務切換器中顯示該應用的內容預覽\n\n重新啟動應用以生效此更改</string>
 
 </resources>
