+++ app/src/main/res/xml/about_settings.xml
@@ -72,18 +72,6 @@
         android:summary="@string/settings_about_beta_update_channel_summary"
         android:defaultValue="@bool/beta"/>
 
-    <com.hippo.preference.SwitchPreference
-        android:key="enable_analytics"
-        android:title="@string/settings_about_enable_analytics"
-        android:defaultValue="true"/>
-
-    <com.hippo.preference.MessagePreference
-        android:key="about_analytics"
-        android:title="@string/settings_about_about_analytics"
-        app:dialogMessage="@string/analytics_explain"
-        app:dialogMessageHtml="true"
-        app:dialogMessageLinkify="true"/>
-
     <Preference
         android:key="donate"
         android:title="@string/settings_about_donate"/>
